taiwan_stock_app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
requests==2.31.0
pandas==2.2.0
google-generativeai==0.8.3
websockets==12.0
apscheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
yfinance>=0.2.36

# Google OAuth
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
